08_SingleAgent/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
openai
chromadb
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
requests
beautifulsoup4
python-dotenv
pdfplumber
pymupdf
